further

+++ src/components/Expresssion/index.tsx
@@ -6,18 +6,21 @@ const Expression = () => {
 const { dispatch } = useContext(ItemContext);
 const { items } = useContext(ItemContext).state;
 const doubleClickFn = (droppableId: string, index: number) => {
+  if (items[index].itemType !== 'condition') {
     dispatch({
       type: 'delete',
       payload: {index}
     })
-  }
+  }}
+  
 return (
 <MyDroppable
 changeOnRightClick={true}
 doubleClickFn={doubleClickFn}
+direction={'horizontal'}
 droppableId={'first'}
-header={`drop`}
-height={800}
+header={`Your Expression`}
+height={400}
 items={items}
 />
 )}
